Refactor: some more renames
